fix: 🐛 fix vector(n)时的push_back()

+++ src/ImgProcessor.hpp
@@ -172,7 +172,7 @@ public:
      * @param[in] iSrc 输入图像
      * @param[out] oDst 输出图像
      * @param[in] iFilterSize 滤波器的 size
-     * @param[in] idD alpha值
+     * @param[in] idD 邻域 S_{xy} 内删除 g(r,c) 的 d 个最低灰度值和 d 个最高灰度值
      */
-    static void ModifiedAlphaMeanFilter(cv::Mat const &iSrc, cv::Mat &oDst, cv::Size iFilterSize, double idD);
+    static void ModifiedAlphaMeanFilter(cv::Mat const &iSrc, cv::Mat &oDst, cv::Size iFilterSize, int idD);
 };
